+++ chevah/utils/logger.py
@@ -136,10 +136,9 @@ class StdOutHandler(StreamHandler, object):
     Logs are not persisted.
     """
 
-    _name = 'Standard output'
-
     def __init__(self):
         super(StdOutHandler, self).__init__()
+        self.name = 'Standard output'
 
     def flush(self):
         '''Flushes the stream.'''
@@ -180,8 +179,6 @@ if os.name == 'nt':
 
         For now we don't use any special category or level.
         """
-
-        _name = 'Windows Eventlog'
 
         def getMessageID(self, log_entry):
             try:
@@ -317,13 +314,14 @@ class _Logger(ObserverMixin):
 
         Return the added handler, or `None.
         """
-        if not self._configuration.syslog:
+        syslog = self._configuration.syslog
+        if not syslog:
             return None
 
         try:
             handler = SysLogHandler(
-                self._configuration.syslog, facility=SysLogHandler.LOG_DAEMON)
-            handler._name = 'Syslog'
+                syslog, facility=SysLogHandler.LOG_DAEMON)
+            handler.name = u'Syslog at %s' % str(syslog)
         except Exception, error:
             raise UtilsError(u'1013',
                 _(u'Failed to start the Syslog logger. %s' % (error)))
@@ -352,6 +350,7 @@ class _Logger(ObserverMixin):
                 dllname=self.NT_EVENTLOG_DDL,
                 logtype=self.NT_EVENTLOG_TYPE,
                 )
+            handler.name = 'Windows Event as %s' % (source_name)
         except Exception, error:
             raise UtilsError(u'1014',
                 _(u'Failed to start the Windows Event logger. %s' % (error)))
@@ -379,6 +378,8 @@ class _Logger(ObserverMixin):
             if self._configuration.file_rotate_external:
                 handler = WatchedFileHandler(
                     log_path, encoding='utf-8')
+                handler.name = u'External rotated file %s' % (
+                    self._configuration.file)
             elif each and each[0] > 0:
                 interval_count, interval_type = each
                 handler = TimedRotatingFileHandler(
@@ -388,6 +389,8 @@ class _Logger(ObserverMixin):
                     backupCount=count,
                     encoding='utf-8',
                     )
+                handler.name = u'Time base rotated file %s at %s' % (
+                    self._configuration.file, each)
             elif bytes:
                 handler = RotatingFileHandler(
                     log_path,
@@ -395,10 +398,12 @@ class _Logger(ObserverMixin):
                     backupCount=count,
                     encoding='utf-8',
                     )
+                handler.name = u'Size base rotated file %s at %s bytes' % (
+                    self._configuration.file, bytes)
             else:
                 handler = FileHandler(log_path, encoding='utf-8')
+                handler.name = u'File %s' % (self._configuration.file)
 
-            handler._name = 'File'
             self.addHandler(handler, patch_format=True)
         except Exception, error:
             raise UtilsError(u'1010',
@@ -479,7 +484,7 @@ class _Logger(ObserverMixin):
 
         self._log.addHandler(handler)
         self._new_handler_added = True
-        if handler._name:
+        if handler.name:
             self.notify('add-handler', Signal(self, name=handler._name))
 
     def removeHandler(self, handler):
@@ -490,7 +495,7 @@ class _Logger(ObserverMixin):
             return
         self._log.removeHandler(handler)
         handler.close()
-        if handler._name:
+        if handler.name:
             self.notify('remove-handler', Signal(self, name=handler._name))
 
     def removeAllHandlers(self):